+++ commons/gridgo-bean/src/main/java/io/gridgo/bean/BReference.java
@@ -26,9 +26,10 @@ public interface BReference extends BElement {
     }
 
     /**
-     * return an instanceof BReference which deserialized from input stream. if stream deserialized to key-value,
-     * it use BObject.toPojo() method. if stream deserialized to breference which contains an instanceof
-     * targetType, return itself.
+     * return an instanceof BReference which deserialized from input stream. if
+     * stream deserialized to key-value, it use BObject.toPojo() method. if stream
+     * deserialized to breference which contains an instanceof targetType, return
+     * itself.
      *
      * @param inputStream    input stream
      * @param serializerName serialzier name, null mean default (msgpack)
@@ -45,7 +46,8 @@ public interface BReference extends BElement {
         if (element.isObject())
             return BFactory.DEFAULT.newReference(element.asObject().toPojo(targetType));
 
-        throw new BeanSerializationException("Cannot convert input bytes as BReference of '" + targetType + "', deserialized: " + element);
+        throw new BeanSerializationException(
+                "Cannot convert input bytes as BReference of '" + targetType + "', deserialized: " + element);
     }
 
     static BReference ofBytes(byte[] bytes, String serializerName, Class<?> targetType) {
@@ -54,26 +56,6 @@ public interface BReference extends BElement {
 
     static BReference ofBytes(ByteBuffer buffer, String serializerName, Class<?> targetType) {
         return ofBytes(new ByteBufferInputStream(buffer), serializerName, targetType);
-    }
-
-    @Override
-    default boolean isContainer() {
-        return false;
-    }
-
-    @Override
-    default boolean isArray() {
-        return false;
-    }
-
-    @Override
-    default boolean isValue() {
-        return false;
-    }
-
-    @Override
-    default boolean isObject() {
-        return false;
     }
 
     @Override
@@ -121,7 +103,8 @@ public interface BReference extends BElement {
     }
 
     /**
-     * Support for I/O operator when reference object is ByteBuffer or InputStream or File
+     * Support for I/O operator when reference object is ByteBuffer or InputStream
+     * or File
      *
      * @param output where data will be write to
      * @return success or not
