--- commons/gridgo-bean/src/main/java/io/gridgo/bean/SerializableReference.java
@@ -1,8 +1,0 @@
-package io.gridgo.bean;
-
-public interface SerializableReference {
-
-    String toXml();
-
-    String toJson();
-}
