+++ commons/gridgo-bean/src/main/java/io/gridgo/bean/impl/AbstractBContainer.java
@@ -2,6 +2,7 @@ package io.gridgo.bean.impl;
 
 import io.gridgo.bean.BContainer;
 import io.gridgo.bean.factory.BFactory;
+import io.gridgo.utils.annotations.Transient;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,5 +10,6 @@ public abstract class AbstractBContainer extends AbstractBElement implements BCo
 
     @Setter
     @Getter
-    private BFactory factory;
+    @Transient
+    private transient BFactory factory;
 }
