+++ commons/gridgo-bean/src/main/java/io/gridgo/bean/impl/AbstractBElement.java
@@ -2,6 +2,7 @@ package io.gridgo.bean.impl;
 
 import io.gridgo.bean.BElement;
 import io.gridgo.bean.serialization.BSerializerRegistry;
+import io.gridgo.utils.annotations.Transient;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,5 +10,6 @@ public abstract class AbstractBElement implements BElement {
 
     @Setter
     @Getter
-    private BSerializerRegistry serializerRegistry;
+    @Transient
+    private transient BSerializerRegistry serializerRegistry;
 }
